reorganiza arquivos do provider jupiter
