+++ trader/providers/jupiter/jupiter_data.py
